refactored ballotProtocol out of Slot
